# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Email:
    properties:
      bcc:
        type: string
      bijlagen_count:
        type: integer
      categorie:
        type: string
      cc:
        type: string
      created_at:
        type: string
      datum:
        type: string
      from_addr:
        type: string
      gmail_id:
        type: string
      heeft_bijlagen:
        type: boolean
      id:
        type: string
      is_draft:
        type: boolean
      is_gelezen:
        type: boolean
      is_ster:
        type: boolean
      is_verwijderd:
        type: boolean
      label_ids:
        items:
          type: string
        type: array
      ontvangen:
        type: integer
      search_text:
        type: string
      snippet:
        type: string
      subject:
        type: string
      synced_at:
        type: string
      thread_id:
        type: string
      to_addr:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Backend REST API for Jeffries Homeapp.
  title: Jeffries Homeapp API
  version: "1.0"
paths:
...
  /emails:
    get:
      description: Returns paginated emails for the user, optionally filtered by category
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - default: 50
        description: Limit count
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset count
        in: query
        name: offset
        type: integer
      - description: Category filter
        in: query
        name: categorie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Email'
            type: array
        "400":
          description: user_id is required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List emails
      tags:
      - Emails
  /emails/delete:
    patch:
      consumes:
      - application/json
      description: Soft-deletes a specific email
      parameters:
      - description: Delete Details (user_id, gmail_id)
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: status ok
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request body
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete email
      tags:
      - Emails
  /emails/read:
    patch:
      consumes:
      - application/json
      description: Updates the read status of a specific email
      parameters:
      - description: Read Status Details (user_id, gmail_id, read)
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: status ok
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request body
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Mark email read status
      tags:
      - Emails
  /emails/search:
    get:
      description: Performs a full-text search across user emails
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 50
        description: Limit count
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Email'
            type: array
        "400":
          description: user_id and q are required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search emails
      tags:
      - Emails
  /emails/stats:
    get:
      description: Returns total and unread email counts for the user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: user_id is required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get email stats
      tags:
      - Emails
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"